+++ sklearnex/dispatcher.py
@@ -18,10 +18,18 @@
 import os
 import sys
 from functools import lru_cache
+from types import ModuleType
 from typing import Optional, Union
 
 from daal4py.sklearn._utils import daal_check_version
-from daal4py.sklearn.monkeypatch.dispatcher import PatchMap
+
+# dict key: sklearn name
+# dict value: tuple entries:
+# - module from sklearn
+# - name of class/function within module
+# - sklearnex/daal4py replacement
+# - sklearn original if any
+PatchMap = dict[str, tuple[ModuleType, str, object, Optional[object]]]
 
 
 def _is_preview_enabled() -> bool:
@@ -398,6 +406,48 @@ def get_patch_names() -> list[str]:
     return list(get_patch_map().keys())
 
 
+# Comment 2026-08-17: These functions were taken from a previous daal4py
+# patching codebase, which originally had more options and arguments for
+# functions. Some calling chains might have redundant steps and could be
+# simplified further.
+def do_patch(name: str, map: PatchMap) -> None:
+    descriptor = map.get(name)
+    if descriptor is None:
+        raise ValueError("Has no patch for: " + name)
+    which, what, replacer, _ = descriptor
+    setattr(which, what, replacer)
+
+
+def do_unpatch(name: str, map: PatchMap) -> None:
+    descriptor = map.get(name)
+    if descriptor is None:
+        raise ValueError("Has no patch for: " + name)
+    which, what, _, replacer = descriptor
+    if replacer is not None:
+        setattr(which, what, replacer)
+    elif hasattr(which, what):
+        delattr(which, what)
+
+
+def enable(
+    name: Optional[str] = None,
+    map: Optional[PatchMap] = None,
+):
+    if name is not None:
+        do_patch(name, map=map)
+    else:
+        for key in map:
+            do_patch(key, map=map)
+
+
+def disable(name: Optional[str] = None, map: Optional[PatchMap] = None):
+    if name is not None:
+        do_unpatch(name, map=map)
+    else:
+        for key in map:
+            do_unpatch(key, map=map)
+
+
 def patch_sklearn(
     name: Optional[Union[str, list[str]]] = None,
     verbose: bool = True,
@@ -475,8 +525,6 @@ def patch_sklearn(
 
         patch_sklearn_global(name, verbose)
 
-    from daal4py.sklearn import patch_sklearn as patch_sklearn_orig
-
     patch_map: PatchMap = get_patch_map()
 
     if name is not None:
@@ -488,14 +536,12 @@ def patch_sklearn(
             "sklearn.utils.parallel._funcwrapper",
         ]
         for name_mandatory in names_mandatory:
-            patch_sklearn_orig(
-                name_mandatory, verbose=False, deprecation=False, map=patch_map
-            )
+            enable(name_mandatory, map=patch_map)
     if isinstance(name, list):
         for algorithm in name:
-            patch_sklearn_orig(algorithm, verbose=False, deprecation=False, map=patch_map)
+            enable(algorithm, map=patch_map)
     else:
-        patch_sklearn_orig(name, verbose=False, deprecation=False, map=patch_map)
+        enable(name, map=patch_map)
 
     if verbose and sys.stderr is not None:
         sys.stderr.write(
@@ -530,17 +576,43 @@ def unpatch_sklearn(
         from sklearnex.glob.dispatcher import unpatch_sklearn_global
 
         unpatch_sklearn_global()
-    from daal4py.sklearn import unpatch_sklearn as unpatch_sklearn_orig
 
     patch_map: PatchMap = get_patch_map()
 
     if isinstance(name, list):
         for algorithm in name:
-            unpatch_sklearn_orig(algorithm, map=patch_map)
+            disable(algorithm, map=patch_map)
     else:
-        unpatch_sklearn_orig(name, map=patch_map)
+        disable(name, map=patch_map)
     if os.environ.get("SKLEARNEX_PREVIEW") == "enabled_via_patch_sklearn":
         os.environ.pop("SKLEARNEX_PREVIEW")
+
+
+def check_is_enabled(name: str, map: PatchMap) -> bool:
+    descriptor = map.get(name)
+    if descriptor is None:
+        return False
+    which, what, replacer, _ = descriptor
+    current = getattr(which, what, None)
+    if current is None:
+        return False
+    return current == replacer
+
+
+def check_patch_is_enabled(
+    name: Optional[str] = None, return_map: bool = False, *, map: PatchMap
+) -> Union[bool, dict[str, bool]]:
+    if name is not None:
+        return check_is_enabled(name, map=map)
+    if return_map:
+        enabled = {}
+        for key in map:
+            enabled[key] = check_is_enabled(key, map=map)
+    else:
+        enabled = True
+        for key in map:
+            enabled = enabled and check_is_enabled(key, map=map)
+    return enabled
 
 
 def sklearn_is_patched(
@@ -569,7 +641,6 @@ def sklearn_is_patched(
     Check : bool or dict[str, bool]
         The patching status of the desired estimators, either as a whole, or
         on a per-estimator basis (output type controlled by ``return_map``)."""
-    from daal4py.sklearn import sklearn_is_patched as sklearn_is_patched_orig
 
     map = get_patch_map()
 
@@ -577,15 +648,15 @@ def sklearn_is_patched(
         if return_map:
             result: dict[str, bool] = {}
             for algorithm in name:
-                result[algorithm] = sklearn_is_patched_orig(algorithm, map=map)
+                result[algorithm] = check_patch_is_enabled(algorithm, map=map)
             return result
         else:
             is_patched = True
             for algorithm in name:
-                is_patched = is_patched and sklearn_is_patched_orig(algorithm, map=map)
+                is_patched = is_patched and check_patch_is_enabled(algorithm, map=map)
             return is_patched
     else:
-        return sklearn_is_patched_orig(name, return_map=return_map, map=map)
+        return check_patch_is_enabled(name, return_map=return_map, map=map)
 
 
 def is_patched_instance(instance: object) -> bool:
